refactoring packages and class names
