+++ GitIssuesApp/app/src/main/java/com/leme/gitissuesapp/service/MainService.java
@@ -1,7 +1,8 @@
-package com.leme.gitissuesapp.main_activity;
+package com.leme.gitissuesapp.service;
 
 import com.leme.gitissuesapp.api.ApiClient;
 import com.leme.gitissuesapp.api.ApiInterface;
+import com.leme.gitissuesapp.contract.MainContract;
 import com.leme.gitissuesapp.model.Issues;
 
 import java.util.List;
@@ -10,10 +11,10 @@ import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
 
-public class MainModel implements MainContract.Model {
+public class MainService implements MainContract.Service {
 
     @Override
-    public void getIssuesList(final OnFinishListener onFinishListener) {
+    public void getIssues(final RequestListener requestListener) {
 
         ApiInterface api = ApiClient.getClient().create(ApiInterface.class);
 
@@ -23,14 +24,14 @@ public class MainModel implements MainContract.Model {
             public void onResponse(Call<List<Issues>> call, Response<List<Issues>> response) {
 
                 List<Issues> issuesList = response.body();
-                onFinishListener.onFinished(issuesList);
+                requestListener.success(issuesList);
 
             }
 
             @Override
             public void onFailure(Call<List<Issues>> call, Throwable throwable) {
 
-                onFinishListener.onFailure(throwable);
+                requestListener.error(throwable);
 
             }
         });
