+++ GitIssuesApp/app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".main_activity.MainActivity">
+    tools:context=".view.activity.MainActivity">
 
     <TextView
         android:id="@+id/hello"
